Changed class names
Following Microsoft naming rules
