+++ TicTacToe/Debug/TicTacToe.tlog/TicTacToe.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
-Debug|Win32|D:\C++\TicTacToe\|
+Debug|Win32|D:\C++\Github\C-with-Kyo\TicTacToe\|
