+++ TicTacToe/TicTacToe.vcxproj
@@ -148,12 +148,12 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="player.cpp" />
-    <ClCompile Include="checking.cpp" />
+    <ClCompile Include="Checker.cpp" />
     <ClCompile Include="TicTacToe.cpp" />
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="checking.hpp" />
+    <ClInclude Include="Checker.hpp" />
     <ClInclude Include="player.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
