+++ TicTacToe/main.cpp
@@ -1,9 +1,16 @@
 #include <iostream>
+#include <string>
 #include "TicTacToe.cpp"
+#include "Checker.hpp"
+#include "Player.hpp"
+
+using namespace std;
+
+string getUserName(string player) {
+    
+}
 
 int main()
 {
-    string getUserName(int player) {
-        string player = "Player";
-    }
+    
 }
